+++ src/App.js
@@ -1,16 +1,25 @@
 import React from 'react'
 import './App.css'
+import { BrowserRouter as Router, Route } from "react-router-dom"
 
 import Header from './Header'
-import TacoGenerator from './TacoGenerator'
+import LinkGenerator from './LinkGenerator'
+import RandomTaco from './RandomTaco'
+import RandomDrink from './RandomDrink'
+// import RandomDrink from './RandomDrink'
+
 
 
 function App() {
   return (
-    <div>
-      <Header />
-      <TacoGenerator />
-    </div>
+    <Router>
+      <div>
+        <Header />
+        <LinkGenerator />
+       <Route path="/randomtaco" component={RandomTaco} />
+       <Route path="/randomdrink" component={RandomDrink} />
+     </div>
+    </Router>
   )
 }
